backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
requests==2.31.0
groq==0.4.2
python-socketio==5.11.1
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
pyjwt==2.8.0
websocket-client==1.7.0
smartapi-python
pyotp
logzero
pycryptodome
<<<<<<< HEAD
google-cloud-aiplatform
google-generativeai
=======
paypalrestsdk==1.13.1 
>>>>>>> 0ad71ac38ac4756de4629f059fd933e2dab09280
